Extract SoruceGenerator common functions This prepares the source generator code for the Gtk-4.0.Integration project. It extracts common functions and places the "GeneratedCode" attribute on the members instead of the class.
