+++ src/Extensions/GObject-2.0.Integration/SourceGenerator/Subclass/HandleAttribute.cs
