+++ src/Extensions/GObject-2.0.Integration/SourceGenerator/Subclass/SubclassValuesProvider.cs
@@ -1,6 +1,4 @@
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using Microsoft.CodeAnalysis;
 
@@ -34,11 +32,11 @@ internal static class SubclassValuesProvider
         if (parentHandle is null)
             return null;
 
-        var accessibility = GetAccessibility(context.TargetSymbol);
+        var accessibility = subclass.GetDeclaredAccessibilityAsString();
         if (accessibility is null)
             return null;
 
-        var upperNestedClasses = GetUpperNestedClasses(context.TargetSymbol);
+        var upperNestedClasses = subclass.GetUpperNestedClasses();
         if (upperNestedClasses is null)
             return null;
 
@@ -50,82 +48,9 @@ internal static class SubclassValuesProvider
             Namespace: context.TargetSymbol.ContainingNamespace.ToDisplayString(),
             IsGlobalNamespace: context.TargetSymbol.ContainingNamespace.IsGlobalNamespace,
             Accessibility: accessibility,
-            FileName: GetFileName(subclass),
+            FileName: subclass.GetFileName(),
             UpperNestedClasses: upperNestedClasses
         );
-    }
-
-    private static Stack<TypeData>? GetUpperNestedClasses(ISymbol symbol)
-    {
-        var stack = new Stack<TypeData>();
-        var containingType = symbol.ContainingType;
-
-        while (containingType is not null)
-        {
-            var accessibility = GetAccessibility(containingType);
-            if (accessibility is null)
-                return null;
-
-            var kind = GetKind(containingType);
-            if (kind is null)
-                return null;
-
-            var typeData = new TypeData(
-                NameGenericArguments: containingType.ToDisplayString(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                Accessibility: accessibility,
-                Kind: kind
-            );
-            stack.Push(typeData);
-            containingType = containingType.ContainingType;
-        }
-
-        return stack;
-    }
-
-    private static string? GetKind(INamedTypeSymbol symbol)
-    {
-        return (symbol.TypeKind, symbol.IsRecord) switch
-        {
-            (TypeKind.Struct, _) => "struct",
-            (TypeKind.Class, true) => "record",
-            (TypeKind.Class, _) => "class",
-            _ => null
-        };
-    }
-
-    private static string GetFileName(INamedTypeSymbol typeSymbol)
-    {
-        var prefix = GetFileNamePrefix(typeSymbol);
-        var suffix = typeSymbol.Arity == 0 ? string.Empty : $"_{typeSymbol.Arity}";
-
-        return prefix + typeSymbol.Name + suffix;
-    }
-
-    private static string GetFileNamePrefix(INamedTypeSymbol typeSymbol)
-    {
-        var sb = new StringBuilder();
-
-        var containingType = typeSymbol.ContainingType;
-        while (containingType is not null)
-        {
-            sb.Insert(0, $"{containingType.Name}.");
-            containingType = containingType.ContainingType;
-        }
-
-        return sb.ToString();
-    }
-
-    private static string? GetAccessibility(ISymbol type)
-    {
-        var accessibility = type.DeclaredAccessibility switch
-        {
-            Accessibility.Public => "public",
-            Accessibility.Internal => "internal",
-            Accessibility.Private => "private",
-            Accessibility.NotApplicable => "internal",
-            _ => null
-        };
-        return accessibility;
     }
 
     private static string? GetParentHandle(ITypeSymbol type)
